+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import * as React from 'react';
 
-import Layout from '@/components/Layout/Layout';
+import Layout from '@/components/Layout';
 import Seo from '@/components/Seo';
 
 export default function NotFoundPage() {
