+++ src/pages/index.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 
 import useHomePage from '@/lib/hooks/useHomePage';
 
+import Layout from '@/components/Layout';
 import Content from '@/components/Layout/Content';
-import Layout from '@/components/Layout/Layout';
 
 function HomePage() {
   useHomePage();
